Inventory items can be used in other inventory items
Updated sprites for bar background, Pato and Lobo.
Added sprites for InventoryContainer and fixed some visual bugs in it.
Added dentures and wolf mask as Props in bar scene.
Created sink Hotspot to use, instead of sink Prop, to go to the sea dream.
Clickables can now be disabled. This make the node invisible and removes
its input detection.
Adding items to inventory and removing them can now yield in order to
run in E.run([]).