Added Space, Spaceship and Planet Homes Plantation
Player is hidden in some scenes to simulate that the player is
controlling another characer.
Created a Spaceship character to travel in Space.
Player can switch between Pato and Lobo in specific scenes.
Added poster Prop to trigger the Space Home daydreaming.
Moved a lot of images to their respective src folders. This is our
attempt to test another project structure.
If the GPI is blocked, adding items to the inventory doesn't trigger the
open and close animations.
A Character can follow the player. THIS HAS TO BE IMPROVED.
Rooms can now change the limits of the camera.