Implemented courage mechanic.
Rooms can save their state in Globals.
One can know if a room was visited, visited for the first time, and
the number of times visited.
Created Amanita, Marrano and Luna planets.
Created Ending scene. Player moves to it after finally talking to Lobo.
Interacting to Lobo changes depending on the level of courage of Pato.
Updated sprites and added lights to Bar scene.

< Closes $216 $217 $226 $227 $228 $229 #22a >