Added puzzl logic for Wolf Mask.
Music in the bar can be changed by talking to Lagarto.
Lagarto behavior changes a bit depending of the state of its mask.
Player can face a clicked Clickable with C.face_clicked().
Added a method to Room to get a Prop by its script_name or name.
Created an introductory narrator dialog to explain the goal of the
game... ojalá y sirva.
--- Ready to work on Conejuno and Vieja chats ---