Created hose-seed pull animation.
Fixed bug of E.run() instructions in queue.
Fixed bug of Cursor.WAIT on dialog options.
Fixed bug of asking for more Golem and Hercules cocktails.
Now characters can talk and move on to the next instruction without
waiting for the player input but an specific time.
Players can't go back to Luna using the same action.
Narrator and Pato reacto to the player trying to go back to Luna using
the same action.